pyproject.toml
[tool.poetry]
name = "kafka-example"
version = "0.1.0"
description = ""
authors = ["timsu92 <33785401+timsu92@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
